chore: move digest from utils to PlanBoardFuncs
