+++ frontend/src/pages/planner/Planner.tsx
@@ -11,14 +11,14 @@ import { Navigate, useParams } from '@tanstack/react-router'
 import { useState, useEffect, useRef, useCallback } from 'react'
 import { type CourseDetails, type Major, DefaultService, type ValidatablePlan, type EquivDetails, type EquivalenceId, type ValidationResult, type PlanView, type CancelablePromise, type ClassId, type CurriculumSpec } from '../../client'
 import { type PseudoCourseDetail, type PseudoCourseId, type CurriculumData, type ModalData, isCourseRequirementErr, type Cyear, type PossibleBlocksList } from './utils/Types'
-import { validateCourseMovement, updateClassesState } from './utils/PlanBoardFunctions'
+import { validateCourseMovement, updateClassesState, locateClassInPlan } from './utils/PlanBoardFunctions'
 import { useAuth } from '../../contexts/auth.context'
 import { toast } from 'react-toastify'
 import DebugGraph from '../../components/DebugGraph'
 import deepEqual from 'fast-deep-equal'
 import { DndProvider } from 'react-dnd'
 import { HTML5Backend } from 'react-dnd-html5-backend'
-import { collectRequirements, handleErrors, PlannerStatus, locateClassInPlan } from './utils/utils'
+import { collectRequirements, handleErrors, PlannerStatus } from './utils/utils'
 import { updateCurriculum, isMinorValid, isMajorValid, loadCurriculumsData } from './utils/CurriculumUtils'
 import ReceivePaste from './utils/ReceivePaste'
 import ModBanner from './ModBanner'
